rename Extension folder
